nvm %zu is the correct argument

+++ clcomm.cpp
@@ -13,7 +13,6 @@
 #include <SDL/SDL.h>
 #include <cstdio>
 #include <cmath>
-#include <stdint.h>
 
 using namespace std;
 
@@ -110,7 +109,7 @@ int main() {
 					for(int j = i; j < clients.size() - 1; j++)
 						clients[j] = clients[j+1];
 					clients.pop_back();
-					printf("Client disconnected size = %lu\n", (uint64_t)clients.size());
+					printf("Client disconnected size = %zu\n", clients.size());
 					i--;
 					break;
 				}
